chore: organize pro dashboard
